Update style.css
Applied this rule: .notion-callout.bg-yellow-light a{background:rgba(35,35,35,.2)!important;color:var(--color-text-light)!important}

Onto the yellow by separating the yellow out into its own line. This is simply a syntactic change; this rule was already being applied to .notion-callout.bg-yellow-light a and a bunch of other coloured callouts using the , operator; this should not result in a change in look or function.